refactor: file moved from a folder into another + fix bugs
